map issue resolved/working render issue
