+++ src/App.js
@@ -1,28 +1,7 @@
 import React from 'react';
 import TodoList from './components/TodoList';
 
-const tasks = [
-  {
-    task: 'Review Lesson of the Day',
-    id: 123,
-    completed: false
-  },
-  {
-    task: 'Complete the Follow Along(s)',
-    id: 1234,
-    completed: false
-  },
-  {
-    task: 'Attend Class Lecture',
-    id: 12345,
-    completed: false
-  },
-  {
-    task: 'Begin Lesson Project',
-    id: 123456,
-    completed: false
-  }
-];
+
 
 class App extends React.Component {
   // you will need a place to store your state in this component.
@@ -32,7 +11,7 @@ class App extends React.Component {
     return (
       <div>
         <h2>Welcome to your Todo App! :-)</h2>
-        <TodoList />
+        <TodoList tasks={this.tasks} />
       </div>
     );
   }
